Moved db to src/ and removed unnecessary environment variable in docker-compose
