+++ docker-compose.yml
@@ -44,8 +44,6 @@ services:
       - "5000:5000"
     env_file:
       - ./backend/.env
-    environment:
-      DB_HOST: db
     volumes:
       - ./backend:/app
       - /app/node_modules
